extract(parser): Extract HCL normalization helpers
- Move HCL config-root normalization into its own file so grammar-specific result handling stays separated from YAML recovery logic
- Rename the YAML result file to reflect its narrower scope while preserving the existing whitespace filtering and body-bound snapping behavior
